git-svn-id: http://red5.googlecode.com/svn/java/server/trunk@1578 1b6495e4-3631-0410-8e05-8f51eee8b9cc

+++ src/org/red5/io/flv/IKeyFrameDataAnalyzer.java
@@ -30,6 +30,6 @@ public interface IKeyFrameDataAnalyzer {
 
 		public int timestamps[];
 
-		public int positions[];
+		public long positions[];
 	}
 }
